+++ qtcore/qrandomgenerator64.go
@@ -97,7 +97,7 @@ func NewQRandomGenerator64(seedValue uint) *QRandomGenerator64 {
 // [-2] void QRandomGenerator64(quint32)
 func NewQRandomGenerator64__() *QRandomGenerator64 {
 	// arg: 0, quint32=Typedef, quint32=Typedef, unsigned int
-	seedValue := 1
+	seedValue := uint(1)
 	rv, err := qtrt.InvokeQtFunc6("_ZN18QRandomGenerator64C2Ej", qtrt.FFI_TYPE_POINTER, seedValue)
 	qtrt.ErrPrint(err, rv)
 	gothis := NewQRandomGenerator64FromPointer(unsafe.Pointer(uintptr(rv)))
